UNO-338 chore: extract common search-widget methods into `AbstractSearchElement`
